don't mutate temporary TargaStream
This commit fixes a compile error generated by zig 0.11.0-dev.3132

Previously we were calling the reader function (Which takes a mutable
pointer) on a temporary TargaStream that was constructed as an argument.

This caused the compile error as all temporary variables are constant.
Defining it as an actual variable allows it to be mutable