jpeg/progressive: Undo decoding single MCUs and then writing to image
In commit e18ed21, Scan was changed to decode & dequantize a single MCU, and then write the to the final pixel destination. This is a bit of a problem for progressive JPEGs which can write to multiple MCUs at once depending on the number of components in the scan (see A.2.2 of ITU-T81).

Additionally this coupled decoding entropy segments, dequantization, and color conversion,  under Scan which made the code base difficult to understand.

Instead place mcu_storage on Frame and have it contain all MCUs for an image (Scan only contained storage for a single MCU). This will increase the memory usage as it was previously but will make progressive JPEG implementation more straightforward.