jpeg: rename/use restart and restart interval markers
Adds parseDefineRestartInterval to store the restart_interval.

Restart intervals (ITU-T81, B.2.4.4) defines the number of MCUs before the decoder should be reset (section E.2.4) which involves resetting DC prediction values to 0 and additionally flushes the remaining bits in the current byte (see E.1.4 which states that the remaining byte is padded with 1s).

After the padded byte, a restart marker is inserted. This is currently ignored.