jpeg: eliminate header struct; move members to scan directly
Deletes the Header struct and moves its members directly to scan. This is done for two reasons, first for progressive scans we'll want to do error checking against the frame_type which is not currently accessible to Header. We could instead additionally pass frame_type as an additional parameter but since the only purpose of the Header struct is to pass it to the Scan, just inline the members.