jpeg: add readBits; add assert to consume bits
Add convenience function readBits which will call peekBits and consumeBits. Add an assert to consumeBits. Consume bits will no longer return an error if num_bits > 16, it is now a programmer error to call consumeBits with more bits than it has in the buffer. The programmer can check bit_count to determine the number of bits or peekBits to fill the buffer. For flushing, call flushBits.