jpeg: fix images with restart interval
With restart intervals, when the correct number of MCUs are decoded, we're supposed to flush the remaining bits of the current byte. Previously we just emptied the bit buffer. However since we're now speculatively filling the bit buffer, there may be its that shouldn't be flushed. Instead determine if we're in the middle of a byte and flush rest of the bits in that byte.

Since restart interval updates can occur between scans, we can't initialize it onto the Frame which is not modifiable. Pass it to the Scan instead.