jpeg: assorted bug fixes for progressive; add test
jpeg.zig: progressive jpegs have multiple scans and thus pixels are initialized multiple times. Instead initialize once on the frame.

Scan.zig: restart intervals are passed between scans and thus can't live on the frame (which is const) store it on jpeg and pass it directly to Scan.

Add test for restart and progressive in the same image.