eliminate usage of `usingnamespace`
`usingnamespace` will likely be removed from the language in 0.15 or
0.16, and is already incompatible with incremental compilation, which
adds a lot of friction to the development process when using `zigimg` as
a dependency in a larger project.

This commit replaces the various mixin methods with namespaces for `to`
and `from` methods. Unfortunately, indirect decl literals don't seem to
be supported so `const c: SomeColor = .from.rgb(...)` does not work :/

This commit makes the API different between rgba colors and others such
as grayscale, this should probably be addressed in the future by making
similar namespaces for other color types.

As a bonus, in the process of making these namespaces I did end up
unifying a lot of logic; I'm particularly proud of `ScaleValue`.